Added Windows compatibility. Still WIP.

+++ core/libs/shared/osOps.py
@@ -1,38 +1,60 @@
 #!/usr/bin/python
-#support	:Nuno Pereira - nuno.pereira@gps-ldn.com
-#title     	:dirOps
-#copyright	:Gramercy Park Studios
+#support    :Nuno Pereira - nuno.pereira@gps-ldn.com
+#title      :dirOps
+#copyright  :Gramercy Park Studios
 
+# Manages OS operations
 
-#manages OS operations
 import os, re
 
 #creates directory for the specified path with the specified umask
 def createDir(path, umask='000'):
 	if not os.path.isdir(path):
-		os.system('%s; mkdir -p %s' % (setUmask(umask), path))
+		if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+			os.makedirs(path)
+		else:
+			os.system('%s; mkdir -p %s' % (setUmask(umask), path))
 		return path
 
 #Sets permissions to provided path
 def setPermissions(path, mode='a+w'):
-	os.system('chmod -R %s %s' % (mode, path))
+	if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+		os.chmod(path, 0o777)
+	else:
+		os.system('chmod -R %s %s' % (mode, path))
 	return path
 
 #hardlinks files with the set umask
 def hardLink(source, destination, umask='000'):
-	os.system('%s; ln -f %s %s' % (setUmask(umask), source, destination))
+	if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+		os.system('mklink /H %s %s' % (destination, source))
+	else:
+		os.system('%s; ln -f %s %s' % (setUmask(umask), source, destination))
 	return destination
 
 #removes files or folders recursively
 def recurseRemove(path):
-	os.system('rm -rf %s' % path)
+	if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+		if os.path.isdir(path):
+			os.system('rmdir %s /s /q' % path)
+		else:
+			os.system('del %s /f /q' % path)			
+	else:
+		os.system('rm -rf %s' % path)
 	return path
 
+#copy the contents of a folder recursively - rewrite using shutil.copy / copytree
 def copyDirContents(source, destination, umask='000'):
-	os.system('%s; cp -rf %s %s' % (setUmask(umask), os.path.join(source, '*'), destination))
+	if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+		os.system('copy %s %s' %(os.path.join(source, '*'), destination))
+	else:
+		os.system('%s; cp -rf %s %s' % (setUmask(umask), os.path.join(source, '*'), destination))
 
 def setUmask(umask='000'):
-	return 'umask %s' % umask
+	if os.environ['ICARUS_RUNNING_OS'] == 'Windows':
+		return ""
+	else:
+		return 'umask %s' % umask
 
 #sanitizes characters in string. Default replaces all non-alphanumeric characters with nothing.
 def sanitize(instr, pattern='\W', replace=''):
